added funtion send_work_time
